% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_valid_protein_sequences.R
\name{are_valid_protein_sequences}
\alias{are_valid_protein_sequences}
\title{Determine if these are all valid protein sequences}
\usage{
are_valid_protein_sequences(protein_sequences, verbose = FALSE)
}
\arguments{
\item{protein_sequences}{one ore more protein sequence,
each sequence with the amino acids as capitals, for
example \code{MEILCEDNTSLSSIPNSL}}

\item{verbose}{set to TRUE for more output}
}
\value{
TRUE if the protein sequence is valid
}
\description{
Determine if these are all valid protein sequences,
as can be used in topology prediction
}
