% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_distance_to_tmh_center_from_topology_str.R
\name{calc_distance_to_tmh_center_from_topology_str}
\alias{calc_distance_to_tmh_center_from_topology_str}
\title{Calculate the the distance for each amino acid to the center
of the TMH}
\usage{
calc_distance_to_tmh_center_from_topology_str(topology_str)
}
\arguments{
\item{topology_str}{the topology as a string,
for example \code{000000111100000}}
}
\value{
a \link[tibble]{tibble} with the columns
'position' and 'distance_to_tmh_center'
}
\description{
Calculate the the distance for each amino acid to the center
of the TMH
}
\author{
Richèl J.C. Bilderbeek
}
