% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_distance_to_tmh_center_from_topology_str_cpp_stl}
\alias{calc_distance_to_tmh_center_from_topology_str_cpp_stl}
\title{Use Rcpp to calculate the distance to a TMH center}
\usage{
calc_distance_to_tmh_center_from_topology_str_cpp_stl(topology_str)
}
\arguments{
\item{topology_str}{a topology as a string}
}
\value{
a vector with distances
}
\description{
Use Rcpp to calculate the distance to a TMH center
}
