% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pureseqtm_installation.R
\name{check_pureseqtm_installation}
\alias{check_pureseqtm_installation}
\title{Checks the installation of PureseqTM.
Throws a helpful error message if incomplete,
else does nothing}
\usage{
check_pureseqtm_installation(folder_name = get_default_pureseqtm_folder())
}
\arguments{
\item{folder_name}{superfolder of PureseqTM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_pureseqtm_folder}}
}
\value{
Nothing
}
\description{
Checks the installation of PureseqTM.
Throws a helpful error message if incomplete,
else does nothing
}
\examples{
  install_pureseqtm()
  check_pureseqtm_installation()
}
\author{
Richèl J.C. Bilderbeek
}
