% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pureseqtm_mode.R
\name{check_pureseqtm_mode}
\alias{check_pureseqtm_mode}
\title{Check if the PureseqTM is valid.
Will \link{stop} if not}
\usage{
check_pureseqtm_mode(mode)
}
\arguments{
\item{mode}{one of the PureseqTM modes:
\itemize{
\item \code{proteome} obtain the topology of one or more proteins
\item \code{protein} obtain detailed information about one protein
}}
}
\value{
Nothing
}
\description{
Check if the PureseqTM is valid.
Will \link{stop} if not
}
