% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_topology.R
\name{check_topology}
\alias{check_topology}
\title{Check if the topology is valid.
Will \link{stop} if not.}
\usage{
check_topology(topology)
}
\arguments{
\item{topology}{the topology as a \link[tibble]{tibble}
as returned by \link{predict_topology}}
}
\description{
Check if the topology is valid.
Will \link{stop} if not.
}
\examples{
library(testthat)

if (is_pureseqtm_installed()) {
  fasta_filename <- get_example_filename("1bhaA.fasta")
  topology <- predict_topology(fasta_filename)
  expect_silent(check_topology(topology))
}
}
\author{
Richèl J.C. Bilderbeek
}
