% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_topology_str.R
\name{check_topology_str}
\alias{check_topology_str}
\title{Check if the topology string is valid.
Will \link{stop} if not.}
\usage{
check_topology_str(topology_str)
}
\arguments{
\item{topology_str}{the topology as a string,
for example \code{000000111100000}}
}
\value{
Nothing. Will \link{stop} with a helpful error message if
the topology is invalid.
}
\description{
Check if the topology string is valid.
Will \link{stop} if not.
}
\examples{
check_topology_str("0000000000000000000000000011111111111111111100000")
}
\author{
Richèl J.C. Bilderbeek
}
