% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_n_tmhs.R
\name{count_n_tmhs}
\alias{count_n_tmhs}
\title{Count the number of TMHs in a topology}
\usage{
count_n_tmhs(topology_str)
}
\arguments{
\item{topology_str}{the topolgy as a string,
for example \code{000000111100000}}
}
\description{
Count the number of TMHs in a topology
}
\examples{
count_n_tmhs("00000000000000000000000000")
count_n_tmhs("00000000001111100000000000")
}
