% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(
  download_url,
  fasta_filename,
  folder_name,
  locatome,
  mode,
  protein_sequence,
  protein_sequences,
  pureseqtm_filename,
  pureseqtm_result,
  temp_folder_name,
  verbose
)
}
\arguments{
\item{download_url}{the URL to download PureseqTM from}

\item{fasta_filename}{path to a FASTA file}

\item{folder_name}{superfolder of PureseqTM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_pureseqtm_folder}}

\item{locatome}{text similar to a FASTA file, with protein names
and the locations of its amino acids}

\item{mode}{one of the PureseqTM modes:
\itemize{
\item \code{proteome} obtain the topology of one or more proteins
\item \code{protein} obtain detailed information about one protein
}}

\item{protein_sequence}{a protein sequence}

\item{protein_sequences}{one or more protein sequences}

\item{pureseqtm_filename}{filename to write the PureseqTM results to}

\item{pureseqtm_result}{the result of a PureseqTM run}

\item{temp_folder_name}{path of a temporary folder.
The folder does not need to exist.
Files that are out in this folder are not automatically
deleted, which is not a problem, as the default
path given by \link{tempdir} is automatically cleaned
by the operating system}

\item{verbose}{set to TRUE for more output}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
\code{@noRd}. This is not done, as this will disallow all
functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
