% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_filenames.R
\name{get_example_filenames}
\alias{get_example_filenames}
\title{Get the full path to all PureseqTM example files}
\usage{
get_example_filenames(folder_name = get_default_pureseqtm_folder())
}
\arguments{
\item{folder_name}{superfolder of PureseqTM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_pureseqtm_folder}}
}
\value{
a character vector with all PureseqTM example files
}
\description{
Get the full path to all PureseqTM example files
}
\examples{
library(testthat)

if (is_pureseqtm_installed()) {
  filenames <- get_example_filenames()
  expect_true(all(file.exists(filenames)))
}
}
\seealso{
use \link{get_example_filename} to get the full
path to a PureseqTM example file
}
