% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pureseqtm_example_filename.R
\name{get_pureseqtm_example_filename}
\alias{get_pureseqtm_example_filename}
\title{Get the full path to a PureseqTM example files}
\usage{
get_pureseqtm_example_filename(
  filename,
  folder_name = get_default_pureseqtm_folder()
)
}
\arguments{
\item{filename}{name of the example file, without the path}

\item{folder_name}{superfolder of PureseqTM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_pureseqtm_folder}}
}
\description{
Get the full path to a PureseqTM example files
}
\examples{
library(testthat)

expect_true(file.exists(get_pureseqtm_example_filename("1bhaA.fasta")))
expect_true(file.exists(get_pureseqtm_example_filename("human_20416.fasta")))
}
\seealso{
use \link{get_pureseqtm_example_filenames} to get all PureseqTM
example filenames
}
