% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pureseqtm_version.R
\name{get_pureseqtm_version}
\alias{get_pureseqtm_version}
\title{Get the PureseqTM version}
\usage{
get_pureseqtm_version(folder_name = get_default_pureseqtm_folder())
}
\arguments{
\item{folder_name}{superfolder of PureseqTM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_pureseqtm_folder}}
}
\value{
a version number as a character vector of one element,
for example \code{v0.10}
}
\description{
Get the PureseqTM version
}
\examples{
if (is_pureseqtm_installed()) {
  get_pureseqtm_version()
}
}
\author{
Richèl J.C. Bilderbeek
}
