% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pureseqtm.R
\name{install_pureseqtm}
\alias{install_pureseqtm}
\title{Install PureseqTM to a local folder}
\usage{
install_pureseqtm(folder_name = get_default_pureseqtm_folder())
}
\arguments{
\item{folder_name}{superfolder of PureseqTM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_pureseqtm_folder}}
}
\description{
Install PureseqTM to a local folder
}
\examples{
library(testthat)

if (is_on_travis() && !is_pureseqtm_installed()) {
  install_pureseqtm()
  expect_true(is_pureseqtm_installed())
}
}
\author{
Richèl J.C. Bilderbeek
}
