% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_fasta_text.R
\name{is_fasta_text}
\alias{is_fasta_text}
\title{Is the text a FASTA file text?}
\usage{
is_fasta_text(fasta_file_text)
}
\arguments{
\item{fasta_file_text}{text of a FASTA file}
}
\description{
Is the text a FASTA file text?
}
