% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pureseqtm_installed.R
\name{is_pureseqtm_installed}
\alias{is_pureseqtm_installed}
\title{Measure if PureseqTM is installed locally}
\usage{
is_pureseqtm_installed(folder_name = get_default_pureseqtm_folder())
}
\arguments{
\item{folder_name}{superfolder of PureseqTM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_pureseqtm_folder}}
}
\value{
\link{TRUE} is PureseqTM is installed locally,
\link{FALSE} otherwise
}
\description{
Measure if PureseqTM is installed locally
}
\examples{
is_pureseqtm_installed()
}
\author{
Richèl J.C. Bilderbeek
}
