% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tmhmm_result.R
\name{is_pureseqtm_result}
\alias{is_pureseqtm_result}
\title{Is the text a locatome text, as would be in a FASTA file?}
\usage{
is_pureseqtm_result(pureseqtm_result)
}
\arguments{
\item{pureseqtm_result}{the result of a PureseqTM run}
}
\description{
Is the text a locatome text, as would be in a FASTA file?
}
