% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tmhmm_set_up.R
\name{is_pureseqtm_set_up}
\alias{is_pureseqtm_set_up}
\title{Determine if PureseqTM is set up}
\usage{
is_pureseqtm_set_up(folder_name = get_default_pureseqtm_folder())
}
\arguments{
\item{folder_name}{superfolder of PureseqTM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_pureseqtm_folder}}
}
\value{
TRUE if PureseqTM is set up, FALSE otherwise
}
\description{
Determine if PureseqTM is set up
}
\author{
Richèl J.C. Bilderbeek
}
