% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_protein_sequence.R
\name{is_valid_protein_sequence}
\alias{is_valid_protein_sequence}
\title{Determine if this a valid protein sequence}
\usage{
is_valid_protein_sequence(protein_sequence, verbose = FALSE)
}
\arguments{
\item{protein_sequence}{a protein sequence, with
the amino acids as capitals, for
example \code{MEILCEDNTSLSSIPNSL}.
Use \link{check_protein_sequence} to check if
a protein sequence is valid.}

\item{verbose}{set to TRUE for more output}
}
\value{
TRUE if the protein sequence is valid
}
\description{
Determine if this is a valid protein sequence,
as can be used in topology prediction
}
