% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fasta_file_as_tibble.R
\name{load_fasta_file_as_tibble}
\alias{load_fasta_file_as_tibble}
\title{Parse a FASTA file to a table
with a \code{name} and \code{sequence} column}
\usage{
load_fasta_file_as_tibble(fasta_filename)
}
\arguments{
\item{fasta_filename}{path to a FASTA file}
}
\value{
a \link[tibble]{tibble}
with a \code{name} and \code{sequence} column
}
\description{
Parse a FASTA file to a table
with a \code{name} and \code{sequence} column
}
\seealso{
use \link{load_fasta_file_as_tibble_cpp}
to directly call the C++ function that does the actual
work. Use \link{load_fasta_file_as_tibble_r} to call
the (approx ten thousand times slower) R function
}
