% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{load_fasta_file_as_tibble_cpp_raw}
\alias{load_fasta_file_as_tibble_cpp_raw}
\title{Do it}
\usage{
load_fasta_file_as_tibble_cpp_raw(fasta_filename)
}
\arguments{
\item{fasta_filename}{FASTA filename}
}
\value{
characters in each element of the vector
}
\description{
Do it
}
