% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fasta_file_as_tibble_r.R
\name{load_fasta_file_as_tibble_r}
\alias{load_fasta_file_as_tibble_r}
\title{Parse a FASTA file to a table
with a \code{name} and \code{sequence} column}
\usage{
load_fasta_file_as_tibble_r(fasta_filename)
}
\arguments{
\item{fasta_filename}{path to a FASTA file}
}
\value{
a \link[tibble]{tibble}
with a \code{name} and \code{sequence} column
}
\description{
Parse a FASTA file to a table
with a \code{name} and \code{sequence} column
}
