% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pureseqtm_proteome_text.R
\name{parse_pureseqtm_proteome_text}
\alias{parse_pureseqtm_proteome_text}
\title{Parse the output of a call to \code{PureseqTM_proteome.sh}}
\usage{
parse_pureseqtm_proteome_text(pureseqtm_proteome_text)
}
\arguments{
\item{pureseqtm_proteome_text}{the output of a call
to \code{PureseqTM_proteome.sh}}
}
\description{
Parse the output of a call to \code{PureseqTM_proteome.sh}
}
