% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_topology.R
\name{predict_topology}
\alias{predict_topology}
\title{Predict the topology of a proteome}
\usage{
predict_topology(
  fasta_filename,
  folder_name = get_default_pureseqtm_folder(),
  topology_filename = tempfile(fileext = ".top")
)
}
\arguments{
\item{fasta_filename}{path to a FASTA file}

\item{folder_name}{superfolder of PureseqTM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_pureseqtm_folder}}

\item{topology_filename}{name of the file to save a protein's
topology to}
}
\value{
a \link[tibble]{tibble}
}
\description{
Predict the topology of a proteome
}
\examples{
library(testthat)

if (is_pureseqtm_installed()) {
  fasta_filename <- get_example_filename("1bhaA.fasta")
  topology <- predict_topology(fasta_filename)
  expect_true("name" \%in\% names(topology))
  expect_true("topology" \%in\% names(topology))
  expect_equal(1, nrow(topology))
}
}
\author{
Richèl J.C. Bilderbeek
}
