% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pureseqtmr_report.R
\name{pureseqtmr_report}
\alias{pureseqtmr_report}
\title{Create a \link{pureseqtmr} report, to be used when reporting bugs}
\usage{
pureseqtmr_report(folder_name = get_default_pureseqtm_folder())
}
\arguments{
\item{folder_name}{superfolder of PureseqTM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_pureseqtm_folder}}
}
\value{
Nothing.
}
\description{
Create a \link{pureseqtmr} report, to be used when reporting bugs
}
\examples{
pureseqtmr_report()
}
\author{
Richèl J.C. Bilderbeek
}
