% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pureseqtm.R
\name{run_pureseqtm}
\alias{run_pureseqtm}
\title{Run PureseqTM}
\usage{
run_pureseqtm(
  fasta_filename,
  folder_name = get_default_pureseqtm_folder(),
  temp_folder_name = tempfile(pattern = "pureseqt_")
)
}
\arguments{
\item{fasta_filename}{path to a FASTA file}

\item{folder_name}{superfolder of PureseqTM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_pureseqtm_folder}}
}
\value{
full path to the files created
}
\description{
Run PureseqTM
}
\examples{
library(testthat)

if (1 == 2 && is_pureseqtm_installed()) {
  fasta_filename <- system.file(
    "extdata", "pureseqtm.fasta", package = "pureseqtm"
  )
  locatome <- run_pureseqtm(fasta_filename)
  cat(locatome, sep = "\n")
}
}
\author{
Richèl J.C. Bilderbeek
}
