% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pureseqtm_proteome.R
\name{run_pureseqtm_proteome}
\alias{run_pureseqtm_proteome}
\title{Run PureseqTM on a proteome}
\usage{
run_pureseqtm_proteome()
}
\value{
the topology of the proteome
}
\description{
Run PureseqTM on a proteome
}
