% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tibble_as_fasta_file.R
\name{save_tibble_as_fasta_file}
\alias{save_tibble_as_fasta_file}
\title{Save the first two columns of a tibble as a FASTA file}
\usage{
save_tibble_as_fasta_file(t, fasta_filename)
}
\arguments{
\item{fasta_filename}{path to a FASTA file}

\item{a}{\link[tibble]{tibble}}
}
\description{
Save the first two columns of a tibble as a FASTA file
}
\author{
Richèl J.C. Bilderbeek
}
