% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_tmhmm.R
\name{set_up_pureseqtm}
\alias{set_up_pureseqtm}
\title{Set up PureseqTM}
\usage{
set_up_pureseqtm(folder_name = get_default_pureseqtm_folder())
}
\arguments{
\item{folder_name}{superfolder of PureseqTM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_pureseqtm_folder}}
}
\description{
Set up PureseqTM
}
\author{
Richèl J.C. Bilderbeek
}
