% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_topology_in_fasta_order.R
\name{sort_topology_in_fasta_order}
\alias{sort_topology_in_fasta_order}
\title{Internal function}
\usage{
sort_topology_in_fasta_order(fasta_filename, t_topology)
}
\arguments{
\item{fasta_filename}{path to a FASTA file}
}
\description{
Sorts the rows of the topology tibble, as returned
by \link{run_pureseqtm_proteome} in the order of names
as the corresponding FASTA file.
}
\details{
PureseqTM does not preserve the order of genes in the FASTA file.
This is bug in PureseqTM, reported at
https://github.com/PureseqTM/PureseqTM_Package/issues/11

Up until then, pureseqtmr needs to correct for this
}
