% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_tmhs.R
\name{tally_tmhs}
\alias{tally_tmhs}
\title{Count the number of transmembrane helices in a topology}
\usage{
tally_tmhs(topology)
}
\arguments{
\item{topology}{the topology as a \link[tibble]{tibble}
with the columns 'name' and 'topology',
where the 'name' column hold all the proteins' names,
and 'topology' contains the respective topologies as strings.}
}
\value{
a \link[tibble]{tibble} with the number of
TMHs per protein
}
\description{
Count the number of transmembrane helices in a topology
}
\examples{
if (is_pureseqtm_installed()) {
  tally_tmhs(
    predict_topology(
      get_example_filename("1bhaA.fasta")
    )
  )
}
}
