% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_tmhmm.R
\name{uninstall_pureseqtm}
\alias{uninstall_pureseqtm}
\title{Uninstall PureseqTM}
\usage{
uninstall_pureseqtm(folder_name = get_default_pureseqtm_folder())
}
\arguments{
\item{folder_name}{name of the folder
where the PureseqTM files are installed.
The name of the PureseqTM binary file will be at
\code{[folder_name]/PureseqTM_Package}}
}
\description{
Uninstall PureseqTM
}
\author{
Richèl J.C. Bilderbeek
}
