% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(
  download_url,
  fasta_filename,
  fasta_file_text,
  folder_name,
  protein_sequence,
  protein_sequences,
  pureseqtm_filename,
  pureseqtm_result,
  pureseqtm_url,
  temp_folder_name,
  topology,
  topology_filename,
  verbose
)
}
\arguments{
\item{download_url}{the URL to download PureseqTM from}

\item{fasta_filename}{path to a FASTA file}

\item{fasta_file_text}{text of a FASTA file}

\item{folder_name}{superfolder of PureseqTM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_pureseqtm_folder}}

\item{protein_sequence}{a protein sequence}

\item{protein_sequences}{one or more protein sequences}

\item{pureseqtm_filename}{filename to write the PureseqTM results to}

\item{pureseqtm_result}{the result of a PureseqTM run}

\item{pureseqtm_url}{URL of the PureseqTM git repository}

\item{temp_folder_name}{path of a temporary folder.
The folder does not need to exist.
Files that are out in this folder are not automatically
deleted, which is not a problem, as the default
path given by \link{tempdir} is automatically cleaned
by the operating system}

\item{topology}{the topology as a \link[tibble]{tibble}
as returned by \link{predict_topology}}

\item{topology_filename}{name of the file to save a protein's
topology to}

\item{verbose}{set to TRUE for more output}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
\code{@noRd}. This is not done, as this will disallow all
functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
