% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pureseqtm_url.R
\name{get_pureseqtm_url}
\alias{get_pureseqtm_url}
\title{Get the URL of the PureseqTM source code}
\usage{
get_pureseqtm_url()
}
\value{
a URL as a character vector of one element
}
\description{
Get the URL of the PureseqTM source code
}
\examples{
get_pureseqtm_url()
}
\author{
Richèl J.C. Bilderbeek
}
