% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{load_fasta_file_as_tibble_cpp_raw}
\alias{load_fasta_file_as_tibble_cpp_raw}
\title{Use Rcpp to load a FASTA file}
\usage{
load_fasta_file_as_tibble_cpp_raw(fasta_filename)
}
\arguments{
\item{fasta_filename}{FASTA filename}
}
\value{
a list with two character vectors,
named 'name' and 'sequence'
}
\description{
Use Rcpp to load a FASTA file
}
