% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_topology.R
\name{plot_topology}
\alias{plot_topology}
\title{Plot the topology}
\usage{
plot_topology(locatome)
}
\arguments{
\item{locatome}{text similar to a FASTA file, with protein names
and the locations of its amino acids}
}
\description{
Plot the topology
}
