% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pureseqtm.R
\name{run_pureseqtm}
\alias{run_pureseqtm}
\title{Runs PureseqTM for one gene and returns the parsed results}
\usage{
run_pureseqtm(
  fasta_filename,
  folder_name = get_default_pureseqtm_folder(),
  temp_folder_name = tempfile(pattern = "pureseqt_")
)
}
\arguments{
\item{fasta_filename}{path to a FASTA file}

\item{folder_name}{superfolder of PureseqTM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_pureseqtm_folder}}

\item{temp_folder_name}{path of a temporary folder.
The folder does not need to exist.
Files that are out in this folder are not automatically
deleted, which is not a problem, as the default
path given by \link{tempdir} is automatically cleaned
by the operating system}
}
\value{
full path to the files created
}
\description{
Runs PureseqTM for one gene and returns the parsed results
}
\examples{
library(testthat)

if (1 == 2 && is_pureseqtm_installed()) {
  fasta_filename <- get_example_filename("1bhaA.fasta")
  topology <- run_pureseqtm(fasta_filename)
  cat(topology, sep = "\n")
}
}
\seealso{
\itemize{
\item{
Use \link{create_pureseqtm_files} to only create the PureseqTM
output files
}
\item{
Use \link{run_pureseqtm_proteome} to run PureseqTM on multiple genes
}
}
}
\author{
Richèl J.C. Bilderbeek
}
