% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tmhmm_to_file.R
\name{run_pureseqtm_to_file}
\alias{run_pureseqtm_to_file}
\title{Creates a FASTA-like file, that has the locations
of the amino acids.}
\usage{
run_pureseqtm_to_file(fasta_filename, pureseqtm_filename)
}
\arguments{
\item{fasta_filename}{path to a FASTA file}

\item{pureseqtm_filename}{filename to write the PureseqTM results to}
}
\description{
Creates a FASTA-like file, that has the locations
of the amino acids.
}
\examples{
library(testthat)

if (1 == 2 && is_pureseqtm_installed()) {
  pureseqtm_filename <- tempfile()
  run_pureseqtm_to_file(
    fasta_filename = system.file("extdata", "pureseqtm.fasta", package = "pureseqtmr"),
    pureseqtm_filename = pureseqtm_filename
  )
  expect_true(file.exists(pureseqtm_filename))
}
}
